+++ packages/examples/cvat/exchange-oracle/alembic/env.py
@@ -23,7 +23,7 @@ if config.config_file_name is not None:
 # from myapp import mymodel
 # target_metadata = mymodel.Base.metadata
 from src.db import Base
-from src.modules.webhook.model import Webhook
+from src.modules.oracle_webhooks.model import Webhook
 from src.modules.cvat.model import Task, Job
 
 target_metadata = Base.metadata
