--- packages/examples/cvat/exchange-oracle/alembic/versions/b819f6138bca_create_webhook_model.py
@@ -1,67 +1,0 @@
-"""create webhook model
-
-Revision ID: b819f6138bca
-Revises: 
-Create Date: 2023-02-28 13:49:22.720696
-
-"""
-from alembic import op
-import sqlalchemy as sa
-import sqlalchemy_utils
-
-
-# revision identifiers, used by Alembic.
-revision = "b819f6138bca"
-down_revision = None
-branch_labels = None
-depends_on = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table(
-        "webhooks",
-        sa.Column("id", sa.String(), nullable=False),
-        sa.Column("signature", sa.String(), nullable=False),
-        sa.Column("escrow_address", sa.String(length=42), nullable=False),
-        sa.Column("network_id", sa.String(), nullable=False),
-        sa.Column(
-            "type",
-            sa.Enum("jl_webhook", "recoracle_webhook", name="webhooktypes"),
-            nullable=True,
-        ),
-        sa.Column(
-            "status",
-            sa.Enum("pending", "completed", name="webhookstatuses"),
-            server_default="pending",
-            nullable=True,
-        ),
-        sa.Column(
-            "created_at",
-            sa.DateTime(timezone=True),
-            server_default=sa.text("now()"),
-            nullable=True,
-        ),
-        sa.Column("updated_at", sa.DateTime(timezone=True), nullable=True),
-        sa.Column(
-            "wait_until",
-            sa.DateTime(timezone=True),
-            server_default=sa.text("now()"),
-            nullable=True,
-        ),
-        sa.PrimaryKeyConstraint("id"),
-        sa.UniqueConstraint("escrow_address"),
-    )
-    op.create_index(op.f("ix_webhooks_id"), "webhooks", ["id"], unique=False)
-    op.create_index(
-        op.f("ix_webhooks_signature"), "webhooks", ["signature"], unique=True
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_index(op.f("ix_webhooks_signature"), table_name="webhooks")
-    op.drop_index(op.f("ix_webhooks_id"), table_name="webhooks")
-    op.drop_table("webhooks")
-    # ### end Alembic commands ###
