+++ packages/examples/cvat/exchange-oracle/src/constants.py
@@ -1,0 +1,7 @@
+from enum import Enum
+from src.config import Config
+
+
+class Networks(str, Enum):
+    polygon_mainnet = Config.polygon_mainnet.network_id
+    polygon_mumbai = Config.polygon_mumbai.network_id
