+++ packages/examples/cvat/exchange-oracle/src/modules/oracle_webhooks/constants.py
@@ -1,10 +1,4 @@
 from enum import Enum
-from src.config import Config
-
-
-class Networks(str, Enum):
-    polygon_mainnet = Config.polygon_mainnet.network_id
-    polygon_mumbai = Config.polygon_mumbai.network_id
 
 
 class WebhookTypes(str, Enum):
