+++ packages/examples/cvat/exchange-oracle/src/validators/validation.py
