+++ packages/examples/cvat/exchange-oracle/tests/api/test_webhook.py
@@ -3,7 +3,7 @@ from fastapi.testclient import TestClient
 
 def test_incoming_webhook_200(client: TestClient) -> None:
     data = {
-        "escrow_address": "0x651D3F1Ac7620eCEc0887200406de58b44854111",
+        "escrow_address": "0xFE776895f6b00AA53969b20119a4777Ed920676a",
         "network": "polygon_mumbai",
     }
     response = client.post(
@@ -27,7 +27,7 @@ def test_incoming_webhook_400(client: TestClient) -> None:
     assert response.status_code == 400
 
     data = {
-        "escrow_address": "0x651D3F1Ac7620eCEc0887200406de58b44854111",
+        "escrow_address": "0xFE776895f6b00AA53969b20119a4777Ed920676a",
         "network": "unsupported_network",
     }
     response = client.post(
